sync code

+++ pkg/registry/core/service/storage/rest.go
@@ -17,6 +17,7 @@ limitations under the License.
 package storage
 
 import (
+	"context"
 	"fmt"
 	"math/rand"
 	"net"
@@ -24,7 +25,6 @@ import (
 	"net/url"
 	"strconv"
 
-	"github.com/golang/glog"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metainternalversion "k8s.io/apimachinery/pkg/apis/meta/internalversion"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -36,13 +36,13 @@ import (
 	"k8s.io/apimachinery/pkg/watch"
 	genericapirequest "k8s.io/apiserver/pkg/endpoints/request"
 	"k8s.io/apiserver/pkg/registry/rest"
-	utilfeature "k8s.io/apiserver/pkg/util/feature"
+	"k8s.io/apiserver/pkg/util/dryrun"
+	"k8s.io/klog"
 
 	apiservice "k8s.io/kubernetes/pkg/api/service"
 	api "k8s.io/kubernetes/pkg/apis/core"
 	"k8s.io/kubernetes/pkg/apis/core/helper"
 	"k8s.io/kubernetes/pkg/apis/core/validation"
-	"k8s.io/kubernetes/pkg/features"
 	registry "k8s.io/kubernetes/pkg/registry/core/service"
 	"k8s.io/kubernetes/pkg/registry/core/service/ipallocator"
 	"k8s.io/kubernetes/pkg/registry/core/service/portallocator"
@@ -69,6 +69,7 @@ type ServiceNodePort struct {
 }
 
 type ServiceStorage interface {
+	rest.Scoper
 	rest.Getter
 	rest.Lister
 	rest.CreaterUpdater
@@ -76,6 +77,7 @@ type ServiceStorage interface {
 	rest.Watcher
 	rest.TableConvertor
 	rest.Exporter
+	rest.StorageVersionProvider
 }
 
 type EndpointsStorage interface {
@@ -107,10 +109,15 @@ func NewREST(
 }
 
 var (
-	_ ServiceStorage          = &REST{}
-	_ rest.CategoriesProvider = &REST{}
-	_ rest.ShortNamesProvider = &REST{}
+	_ ServiceStorage              = &REST{}
+	_ rest.CategoriesProvider     = &REST{}
+	_ rest.ShortNamesProvider     = &REST{}
+	_ rest.StorageVersionProvider = &REST{}
 )
+
+func (rs *REST) StorageVersion() runtime.GroupVersioner {
+	return rs.services.StorageVersion()
+}
 
 // ShortNames implements the ShortNamesProvider interface. Returns a list of short names for a resource.
 func (rs *REST) ShortNames() []string {
@@ -122,6 +129,10 @@ func (rs *REST) Categories() []string {
 	return []string{"all"}
 }
 
+func (rs *REST) NamespaceScoped() bool {
+	return rs.services.NamespaceScoped()
+}
+
 func (rs *REST) New() runtime.Object {
 	return rs.services.New()
 }
@@ -130,23 +141,23 @@ func (rs *REST) NewList() runtime.Object {
 	return rs.services.NewList()
 }
 
-func (rs *REST) Get(ctx genericapirequest.Context, name string, options *metav1.GetOptions) (runtime.Object, error) {
+func (rs *REST) Get(ctx context.Context, name string, options *metav1.GetOptions) (runtime.Object, error) {
 	return rs.services.Get(ctx, name, options)
 }
 
-func (rs *REST) List(ctx genericapirequest.Context, options *metainternalversion.ListOptions) (runtime.Object, error) {
+func (rs *REST) List(ctx context.Context, options *metainternalversion.ListOptions) (runtime.Object, error) {
 	return rs.services.List(ctx, options)
 }
 
-func (rs *REST) Watch(ctx genericapirequest.Context, options *metainternalversion.ListOptions) (watch.Interface, error) {
+func (rs *REST) Watch(ctx context.Context, options *metainternalversion.ListOptions) (watch.Interface, error) {
 	return rs.services.Watch(ctx, options)
 }
 
-func (rs *REST) Export(ctx genericapirequest.Context, name string, opts metav1.ExportOptions) (runtime.Object, error) {
+func (rs *REST) Export(ctx context.Context, name string, opts metav1.ExportOptions) (runtime.Object, error) {
 	return rs.services.Export(ctx, name, opts)
 }
 
-func (rs *REST) Create(ctx genericapirequest.Context, obj runtime.Object, createValidation rest.ValidateObjectFunc, includeUninitialized bool) (runtime.Object, error) {
+func (rs *REST) Create(ctx context.Context, obj runtime.Object, createValidation rest.ValidateObjectFunc, options *metav1.CreateOptions) (runtime.Object, error) {
 	service := obj.(*api.Service)
 
 	if err := rest.BeforeCreate(registry.Strategy, ctx, obj); err != nil {
@@ -164,13 +175,15 @@ func (rs *REST) Create(ctx genericapirequest.Context, obj runtime.Object, create
 	}()
 
 	var err error
-	if service.Spec.Type != api.ServiceTypeExternalName {
-		if releaseServiceIP, err = initClusterIP(service, rs.serviceIPs); err != nil {
-			return nil, err
+	if !dryrun.IsDryRun(options.DryRun) {
+		if service.Spec.Type != api.ServiceTypeExternalName {
+			if releaseServiceIP, err = initClusterIP(service, rs.serviceIPs); err != nil {
+				return nil, err
+			}
 		}
 	}
 
-	nodePortOp := portallocator.StartOperation(rs.serviceNodePorts)
+	nodePortOp := portallocator.StartOperation(rs.serviceNodePorts, dryrun.IsDryRun(options.DryRun))
 	defer nodePortOp.Finish()
 
 	if service.Spec.Type == api.ServiceTypeNodePort || service.Spec.Type == api.ServiceTypeLoadBalancer {
@@ -189,7 +202,7 @@ func (rs *REST) Create(ctx genericapirequest.Context, obj runtime.Object, create
 		return nil, errors.NewInvalid(api.Kind("Service"), service.Name, errs)
 	}
 
-	out, err := rs.services.Create(ctx, service, createValidation, includeUninitialized)
+	out, err := rs.services.Create(ctx, service, createValidation, options)
 	if err != nil {
 		err = rest.CheckGeneratedNameError(registry.Strategy, err, service)
 	}
@@ -207,7 +220,7 @@ func (rs *REST) Create(ctx genericapirequest.Context, obj runtime.Object, create
 	return out, err
 }
 
-func (rs *REST) Delete(ctx genericapirequest.Context, id string, options *metav1.DeleteOptions) (runtime.Object, bool, error) {
+func (rs *REST) Delete(ctx context.Context, id string, options *metav1.DeleteOptions) (runtime.Object, bool, error) {
 	// TODO: handle graceful
 	obj, _, err := rs.services.Delete(ctx, id, options)
 	if err != nil {
@@ -216,13 +229,32 @@ func (rs *REST) Delete(ctx genericapirequest.Context, id string, options *metav1
 
 	svc := obj.(*api.Service)
 
-	// TODO: can leave dangling endpoints, and potentially return incorrect
-	// endpoints if a new service is created with the same name
-	_, _, err = rs.endpoints.Delete(ctx, id, &metav1.DeleteOptions{})
-	if err != nil && !errors.IsNotFound(err) {
-		return nil, false, err
+	// Only perform the cleanup if this is a non-dryrun deletion
+	if !dryrun.IsDryRun(options.DryRun) {
+		// TODO: can leave dangling endpoints, and potentially return incorrect
+		// endpoints if a new service is created with the same name
+		_, _, err = rs.endpoints.Delete(ctx, id, &metav1.DeleteOptions{})
+		if err != nil && !errors.IsNotFound(err) {
+			return nil, false, err
+		}
+
+		rs.releaseAllocatedResources(svc)
 	}
 
+	// TODO: this is duplicated from the generic storage, when this wrapper is fully removed we can drop this
+	details := &metav1.StatusDetails{
+		Name: svc.Name,
+		UID:  svc.UID,
+	}
+	if info, ok := genericapirequest.RequestInfoFrom(ctx); ok {
+		details.Group = info.APIGroup
+		details.Kind = info.Resource // legacy behavior
+	}
+	status := &metav1.Status{Status: metav1.StatusSuccess, Details: details}
+	return status, true, nil
+}
+
+func (rs *REST) releaseAllocatedResources(svc *api.Service) {
 	if helper.IsServiceIPSet(svc) {
 		rs.serviceIPs.Release(net.ParseIP(svc.Spec.ClusterIP))
 	}
@@ -245,18 +277,6 @@ func (rs *REST) Delete(ctx genericapirequest.Context, id string, options *metav1
 			}
 		}
 	}
-
-	// TODO: this is duplicated from the generic storage, when this wrapper is fully removed we can drop this
-	details := &metav1.StatusDetails{
-		Name: svc.Name,
-		UID:  svc.UID,
-	}
-	if info, ok := genericapirequest.RequestInfoFrom(ctx); ok {
-		details.Group = info.APIGroup
-		details.Kind = info.Resource // legacy behavior
-	}
-	status := &metav1.Status{Status: metav1.StatusSuccess, Details: details}
-	return status, true, nil
 }
 
 // externalTrafficPolicyUpdate adjusts ExternalTrafficPolicy during service update if needed.
@@ -292,7 +312,7 @@ func (rs *REST) healthCheckNodePortUpdate(oldService, service *api.Service, node
 	// Allocate a health check node port or attempt to reserve the user-specified one if provided.
 	// Insert health check node port into the service's HealthCheckNodePort field if needed.
 	case !neededHealthCheckNodePort && needsHealthCheckNodePort:
-		glog.Infof("Transition to LoadBalancer type service with ExternalTrafficPolicy=Local")
+		klog.Infof("Transition to LoadBalancer type service with ExternalTrafficPolicy=Local")
 		if err := allocateHealthCheckNodePort(service, nodePortOp); err != nil {
 			return false, errors.NewInternalError(err)
 		}
@@ -300,8 +320,8 @@ func (rs *REST) healthCheckNodePortUpdate(oldService, service *api.Service, node
 	// Case 2: Transition from needs HealthCheckNodePort to don't need HealthCheckNodePort.
 	// Free the existing healthCheckNodePort and clear the HealthCheckNodePort field.
 	case neededHealthCheckNodePort && !needsHealthCheckNodePort:
-		glog.Infof("Transition to non LoadBalancer type service or LoadBalancer type service with ExternalTrafficPolicy=Global")
-		glog.V(4).Infof("Releasing healthCheckNodePort: %d", oldHealthCheckNodePort)
+		klog.Infof("Transition to non LoadBalancer type service or LoadBalancer type service with ExternalTrafficPolicy=Global")
+		klog.V(4).Infof("Releasing healthCheckNodePort: %d", oldHealthCheckNodePort)
 		nodePortOp.ReleaseDeferred(int(oldHealthCheckNodePort))
 		// Clear the HealthCheckNodePort field.
 		service.Spec.HealthCheckNodePort = 0
@@ -310,7 +330,7 @@ func (rs *REST) healthCheckNodePortUpdate(oldService, service *api.Service, node
 	// Reject changing the value of the HealthCheckNodePort field.
 	case neededHealthCheckNodePort && needsHealthCheckNodePort:
 		if oldHealthCheckNodePort != newHealthCheckNodePort {
-			glog.Warningf("Attempt to change value of health check node port DENIED")
+			klog.Warningf("Attempt to change value of health check node port DENIED")
 			fldPath := field.NewPath("spec", "healthCheckNodePort")
 			el := field.ErrorList{field.Invalid(fldPath, newHealthCheckNodePort,
 				"cannot change healthCheckNodePort on loadBalancer service with externalTraffic=Local during update")}
@@ -320,9 +340,21 @@ func (rs *REST) healthCheckNodePortUpdate(oldService, service *api.Service, node
 	return true, nil
 }
 
-func (rs *REST) Update(ctx genericapirequest.Context, name string, objInfo rest.UpdatedObjectInfo, createValidation rest.ValidateObjectFunc, updateValidation rest.ValidateObjectUpdateFunc) (runtime.Object, bool, error) {
+func (rs *REST) Update(ctx context.Context, name string, objInfo rest.UpdatedObjectInfo, createValidation rest.ValidateObjectFunc, updateValidation rest.ValidateObjectUpdateFunc, forceAllowCreate bool, options *metav1.UpdateOptions) (runtime.Object, bool, error) {
 	oldObj, err := rs.services.Get(ctx, name, &metav1.GetOptions{})
 	if err != nil {
+		// Support create on update, if forced to.
+		if forceAllowCreate {
+			obj, err := objInfo.UpdatedObject(ctx, nil)
+			if err != nil {
+				return nil, false, err
+			}
+			createdObj, err := rs.Create(ctx, obj, createValidation, &metav1.CreateOptions{DryRun: options.DryRun})
+			if err != nil {
+				return nil, false, err
+			}
+			return createdObj, true, nil
+		}
 		return nil, false, err
 	}
 	oldService := oldObj.(*api.Service)
@@ -338,9 +370,8 @@ func (rs *REST) Update(ctx genericapirequest.Context, name string, objInfo rest.
 	}
 
 	// Copy over non-user fields
-	// TODO: make this a merge function
-	if errs := validation.ValidateServiceUpdate(service, oldService); len(errs) > 0 {
-		return nil, false, errors.NewInvalid(api.Kind("Service"), service.Name, errs)
+	if err := rest.BeforeUpdate(registry.Strategy, ctx, service, oldService); err != nil {
+		return nil, false, err
 	}
 
 	// TODO: this should probably move to strategy.PrepareForCreate()
@@ -353,19 +384,21 @@ func (rs *REST) Update(ctx genericapirequest.Context, name string, objInfo rest.
 		}
 	}()
 
-	nodePortOp := portallocator.StartOperation(rs.serviceNodePorts)
+	nodePortOp := portallocator.StartOperation(rs.serviceNodePorts, dryrun.IsDryRun(options.DryRun))
 	defer nodePortOp.Finish()
 
-	// Update service from ExternalName to non-ExternalName, should initialize ClusterIP.
-	if oldService.Spec.Type == api.ServiceTypeExternalName && service.Spec.Type != api.ServiceTypeExternalName {
-		if releaseServiceIP, err = initClusterIP(service, rs.serviceIPs); err != nil {
-			return nil, false, err
+	if !dryrun.IsDryRun(options.DryRun) {
+		// Update service from ExternalName to non-ExternalName, should initialize ClusterIP.
+		if oldService.Spec.Type == api.ServiceTypeExternalName && service.Spec.Type != api.ServiceTypeExternalName {
+			if releaseServiceIP, err = initClusterIP(service, rs.serviceIPs); err != nil {
+				return nil, false, err
+			}
 		}
-	}
-	// Update service from non-ExternalName to ExternalName, should release ClusterIP if exists.
-	if oldService.Spec.Type != api.ServiceTypeExternalName && service.Spec.Type == api.ServiceTypeExternalName {
-		if helper.IsServiceIPSet(oldService) {
-			rs.serviceIPs.Release(net.ParseIP(oldService.Spec.ClusterIP))
+		// Update service from non-ExternalName to ExternalName, should release ClusterIP if exists.
+		if oldService.Spec.Type != api.ServiceTypeExternalName && service.Spec.Type == api.ServiceTypeExternalName {
+			if helper.IsServiceIPSet(oldService) {
+				rs.serviceIPs.Release(net.ParseIP(oldService.Spec.ClusterIP))
+			}
 		}
 	}
 	// Update service from NodePort or LoadBalancer to ExternalName or ClusterIP, should release NodePort if exists.
@@ -395,7 +428,7 @@ func (rs *REST) Update(ctx genericapirequest.Context, name string, objInfo rest.
 		return nil, false, errors.NewInvalid(api.Kind("Service"), service.Name, errs)
 	}
 
-	out, created, err := rs.services.Update(ctx, service.Name, rest.DefaultUpdatedObjectInfo(service), createValidation, updateValidation)
+	out, created, err := rs.services.Update(ctx, service.Name, rest.DefaultUpdatedObjectInfo(service), createValidation, updateValidation, forceAllowCreate, options)
 	if err == nil {
 		el := nodePortOp.Commit()
 		if el != nil {
@@ -413,7 +446,7 @@ func (rs *REST) Update(ctx genericapirequest.Context, name string, objInfo rest.
 var _ = rest.Redirector(&REST{})
 
 // ResourceLocation returns a URL to which one can send traffic for the specified service.
-func (rs *REST) ResourceLocation(ctx genericapirequest.Context, id string) (*url.URL, http.RoundTripper, error) {
+func (rs *REST) ResourceLocation(ctx context.Context, id string) (*url.URL, http.RoundTripper, error) {
 	// Allow ID as "svcname", "svcname:port", or "scheme:svcname:port".
 	svcScheme, svcName, portStr, valid := utilnet.SplitSchemeNamePort(id)
 	if !valid {
@@ -466,11 +499,9 @@ func (rs *REST) ResourceLocation(ctx genericapirequest.Context, id string) (*url
 				// but in the expected case we'll only make one.
 				for try := 0; try < len(ss.Addresses); try++ {
 					addr := ss.Addresses[(addrSeed+try)%len(ss.Addresses)]
-					if !utilfeature.DefaultFeatureGate.Enabled(features.ServiceProxyAllowExternalIPs) {
-						if err := isValidAddress(ctx, &addr, rs.pods); err != nil {
-							utilruntime.HandleError(fmt.Errorf("Address %v isn't valid (%v)", addr, err))
-							continue
-						}
+					if err := isValidAddress(ctx, &addr, rs.pods); err != nil {
+						utilruntime.HandleError(fmt.Errorf("Address %v isn't valid (%v)", addr, err))
+						continue
 					}
 					ip := addr.IP
 					port := int(ss.Ports[i].Port)
@@ -486,11 +517,11 @@ func (rs *REST) ResourceLocation(ctx genericapirequest.Context, id string) (*url
 	return nil, nil, errors.NewServiceUnavailable(fmt.Sprintf("no endpoints available for service %q", id))
 }
 
-func (r *REST) ConvertToTable(ctx genericapirequest.Context, object runtime.Object, tableOptions runtime.Object) (*metav1beta1.Table, error) {
+func (r *REST) ConvertToTable(ctx context.Context, object runtime.Object, tableOptions runtime.Object) (*metav1beta1.Table, error) {
 	return r.services.ConvertToTable(ctx, object, tableOptions)
 }
 
-func isValidAddress(ctx genericapirequest.Context, addr *api.EndpointAddress, pods rest.Getter) error {
+func isValidAddress(ctx context.Context, addr *api.EndpointAddress, pods rest.Getter) error {
 	if addr.TargetRef == nil {
 		return fmt.Errorf("Address has no target ref, skipping: %v", addr)
 	}
@@ -558,7 +589,7 @@ func allocateHealthCheckNodePort(service *api.Service, nodePortOp *portallocator
 			return fmt.Errorf("failed to allocate requested HealthCheck NodePort %v: %v",
 				healthCheckNodePort, err)
 		}
-		glog.V(4).Infof("Reserved user requested healthCheckNodePort: %d", healthCheckNodePort)
+		klog.V(4).Infof("Reserved user requested healthCheckNodePort: %d", healthCheckNodePort)
 	} else {
 		// If the request has no health check nodePort specified, allocate any.
 		healthCheckNodePort, err := nodePortOp.AllocateNext()
@@ -566,7 +597,7 @@ func allocateHealthCheckNodePort(service *api.Service, nodePortOp *portallocator
 			return fmt.Errorf("failed to allocate a HealthCheck NodePort %v: %v", healthCheckNodePort, err)
 		}
 		service.Spec.HealthCheckNodePort = int32(healthCheckNodePort)
-		glog.V(4).Infof("Reserved allocated healthCheckNodePort: %d", healthCheckNodePort)
+		klog.V(4).Infof("Reserved allocated healthCheckNodePort: %d", healthCheckNodePort)
 	}
 	return nil
 }
